Add Sentry to upload-service
Sentry requires a configured DSN and environment, which we will
configure directly as part of the systemd service.

This also fixes a bug where region was not being loaded from the config
file. AWS documentation specifies that a flag must be explicitly set to
load region from the ~/aws/config path.

https://docs.aws.amazon.com/sdk-for-javascript/v2/developer-guide/setting-region.html

Issue #9