Create uploader directory during image creation
The uploader directory has survived the removal of the old uploader
service; now we use the uploader directory to stage files that we
download from S3 for processing. Without this directory, the process
daemon tries to download the file from S3, and the S3 SDK throws an
exception because the destination directory does not exist. Previously,
this directory was created by deploying the old uploader service, but
now we need to explicitly create it as part of the AMI creation.

It turns out that a cron job in the task-runner does create this
directory if it does not already exist. As a result, uploads were
failing between a new instance creation and when the hourly cron jobs
run at the :17 minute mark.

Co-authored-by: Andrew Atwood <aatwood@permanent.org>
Co-authored-by: Cecilia Krum <cecilia@permanent.org>
Co-authored-by: Mithuna Krishna <mkrishna@permanent.org>