Deploy notification-service
Deploy the new notification-service[1] to our AWS-hosted
environments. This is very similar to the deployment process for the
upload-service.

This does not yet deploy the notification-service in a working state; in
particular, we have not yet set up the database credentials. However,
this should be enough for us to be able to run the service long enough
to get an error message about the database, which is more than we're
getting now.

Also note that this will temporarily break Vagrant. We had decided not
to put notifications-service into Vagrant[2], due to complications
around how Vagrant affects the node_modules directory[3], particularly
with non-Linux hosts. However, devenv is still very closely tied to this
repo, and separating them further is out of the scope of this work.
Those few of us who are using devenv should not go through a vagrant
destroy/up cycle with this commit until that is resolved.

[1] https://github.com/PermanentOrg/notification-service
[2] https://github.com/PermanentOrg/devenv/pull/19#issuecomment-788132673
[3] https://github.com/PermanentOrg/devenv/issues/12

Issue #45 Add the notification service

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>