Re-format JSON files
Use jq(1) to reformat our JSON files. jq uses two-space indents, puts
each list item on its own line, and puts closing braces on their own
lines.

It sadly doesn't have the ability to edit-in-place, so the process for
each file was something like:

    jq . file.json > file.tmp.json
    mv file.tmp.json file.json

Probably we should set up some kind of automation for this, but for now
we can do it by hand.