Sort JSON object keys
The JSON specification says that the order of keys in objects does not
matter. While there may be an argument that a given order in these files
in particular helps readability, I think that nobody currently working
on this is familiar enough with the code to effectively make that
argument. Instead, enforcing alphabetical order is simple and
unambiguous, and is one fewer thing for us to think about as we work on
this.