Split long Ansible --extra-vars line
Ansible supports variable assignment through the --extra-vars argument,
which must be followed by a single argument which contains a list of
key=value pairs. On the command line, multiple pairs can be combined by
quoting the list; in Packer, the argument is encoded as a JSON string in
the list of arguments.

We have enough arguments that the length of this line is making it
difficult to work with, and it will get worse as we add more variables,
as we intend to.

Fortunately, Ansible will accept multiple instances of the --extra-vars
argument. Having an --extra-vars for each variable is a bit verbose, but
I think in this case it's more readable than the 300+ character line we
had before.