Add Firebase credentials for notification-service
The notification-service needs to talk to Firebase. It expects to
receive credentials through the FIREBASE_CREDENTIALS environment
variable, which should be the minified contents of the JSON file
containing a Google service key.

Create three new GitHub secrets (DEV_NOTIFICATION_FIREBASE_CREDENTIALS,
STAGING_NOTIFICATION_FIREBASE_CREDENTIALS, and
PROD_NOTIFICATION_FIREBASE_CREDENTIALS), and pipe those through to the
systemd service definition template.

(Note that as of this writing, we have not yet created the production
Firebase application; the PROD_NOTIFICATION_FIREBASE_CREDENTIALS GitHub
secret contains placeholder data, and the service is not expected to
start successfully in the prod environment.)