Switch notification.service to EnvironmentFile
As we add more environment variables - and more complex values in those
variables - to the notification-service, the content restrictions of
systemd's `Environment` parameter in the service unit configuration file
become too limiting. In particular, it seems that values containing
double quotes and equal signs cannot be represented in `Environment=`
directives.

There is an alternate way to configure environment variables for systemd
services: the `EnvironmentFile=` directive[1]:

    Similar to Environment= but reads the environment variables from a
    text file. The text file should contain new-line-separated variable
    assignments. Empty lines, lines without an "=" separator, or lines
    starting with ; or # will be ignored, which may be used for
    commenting. A line ending with a backslash will be concatenated with
    the following one, allowing multiline variable definitions. The
    parser strips leading and trailing whitespace from the values of
    assignments, unless you use double quotes (").

There is a security risk to the approach of storing credentials in
environment variables defined in systemd service unit configurations.
From the documentation:

    Note that environment variables are not suitable for passing secrets
    (such as passwords, key material, …) to service processes.
    Environment variables set for a unit are exposed to unprivileged
    clients via D-Bus IPC, and generally not understood as being data
    that requires protection. Moreover, environment variables are
    propagated down the process tree, including across security
    boundaries (such as setuid/setgid executables), and hence might leak
    to processes that should not have access to the secret data.

Specifically, the security risk is that other users and services running
on the same machine are able to retrieve the credentials. Later versions
of systemd add the `LoadCredential=` directive to securely mitigate that
risk, but the version in Debian buster (current stable) does not have
that directive.

In our specific case, I am comfortable accepting this security risk: we
have made the decision to trust the code running on the machine, and the
users who are able to connect to it via ssh. We are limiting the access
we are granting on these credentials. And, finally, it is my hope that
we will not continue with this deployment strategy long term.

[1] https://www.freedesktop.org/software/systemd/man/systemd.exec.html#EnvironmentFile=