Prevent Ansible from mangling the Firebase creds
For reasons that defy understanding, Ansible seems to be intent on
parsing the strings we pass it. Depending on how I attempt to prevent
that behavior, this manifests in a few different ways:

- naive approach: escaped newlines in JSON become actual new lines
- replacing newlines with \n: value is interpreted as JSON/YAML/Python
  dictionary; all quotes are replaced with single quotes
- replacing newlines and calling to_json: the dict is stringified,
  complete with escaped quotes and surrounded with double quotes
- just calling to_json: the dict is still stringified, but this time the
  backslashes in the escaped newlines (\n) in the source data are
  themselves escaped (\\n)

In each one of these cases, the notification-service is not able to
parse the environment variable into the structure it needs.

The solution appears to be to replace the parsed newlines with their
escaped versions, *and* add a leading space to the value to prevent
Ansible from further parsing the result into a dict or whatever. This
results in the file on disk having a leading space in the value, eg
`FIREBASE_CREDENTIALS=' {...`, but JSON doesn't much care about
whitespace and the notification-service should still be able to use this
value to connect to Firebase.