Rename upload-service systemd definition
The language a service is written in is not a useful thing to include in
the name of that service. Even in the historical context in which this
was written, it did not usefully disambiguate between the new
upload-service[1] and the old uploader.

Rename the systemd service definition for the upload-service from
'node-upload.service' to 'upload.service'.

Relatedly, it is helpful to include the organization name in the service
description, so that we may more easily find all the Permanent-related
services in `systemctl list-units` or so.

[1] https://github.com/PermanentOrg/upload-service