Set NODE_ENV=production for microservices
We want to be able to distinguish between production(-esque) and
developer environments in the upload-service. The first motivation for
that is to allow for appropriate logging in each: systemd handles
timestamping log lines automatically, while the normal stdout does not.

The convention for exposing this in Node.js applications is the NODE_ENV
environment variable, which is traditionally set to `production`; in
non-production environments, it is either set to `development` or left
unset.

Set NODE_ENV=production for both the upload-service and the
notification-service.