Enable iOS universal links
Applications on iOS can register themselves as handlers for certain
URLs, but doing so requires coordination and cooporation from the
website. This is done by having the website host a public configuration
at `/.well-known/apple-app-site-association`[1].

Add this file to allow the app to open links to staging.permanent.org.

As we need different values in the different environments (dev, staging,
prod), add a variable to our GitHub Action configuration, and thread it
through to `configure.sh`. See also our earlier notes on how to do
this[2]. Currently we only have a staging id, so use a test value for
dev and leave prod empty. The id isn't a secret, so it's fine to enter
it directly into the image build file.

This is the first well-known file we are hosting; see also the list of
/.well-known/ files[3].

[1] https://developer.apple.com/library/archive/documentation/General/Conceptual/AppSearch/UniversalLinks.html
[2] #51 (comment)
[3] https://en.wikipedia.org/wiki/List_of_/.well-known/_services_offered_by_webservers

VSP-291: [iOS] Backend iOS url deep-linking certificate configuration

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>