Use standard CPU credit mode on cron
The new t3 instance type has different default behavior than the t2
instance type did. Both are burstable instance types, meaning they can
temporarily use more CPU time if needed, and AWS regulates
that with CPU credits. With the t2 instance types, the default was to
use "standard" mode, which throttles the instance if it tries to use too
much CPU for too long; with the t3 instance type, the default changed to
"unlimited" mode, which allows the instance to use as much as it wants
but incurs extra costs.

The work the cron machines do is not very sensitive to delays, and we do
have some wasteful jobs running in prod. While we should fix that, let's
not allow it to waste money in the meantime.

https://aws.amazon.com/about-aws/whats-new/2018/08/introducing-amazon-ec2-t3-instances/
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/burstable-performance-instances-unlimited-mode.html
https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/instance#credit-specification

PER-8256 Upgrade AWS Instance Types