Pass HTTP Authorization header to API
In order for the API to be able to accept OAuth tokens, Apache needs to
pass along the Authorization header[1] clients will use. By default, it
does not; from the documentation on the CGIPassAuth directive[2]:

     Normally these HTTP headers are hidden from scripts. This is to
     disallow scripts from seeing user ids and passwords used to
     access the server when HTTP Basic authentication is enabled in
     the web server.

The web server does not (and will not) use HTTP Basic authentication.
(It may be of historical interest to note that it once did, in order to
protect the dev/staging wordpress sites, but that was functionally
disabled last spring and finally removed in commit
003ab3c60b68c3f4e6f39d5633277eb307e806d3.)

Turn on the CGIPassAuth directive in local, dev, staging, and
production.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Authorization
[2] https://httpd.apache.org/docs/current/en/mod/core.html#cgipassauth

PER-8523 Validate auth tokens on API requests