Update personal access token
When I removed @xmunoz from this organization, her personal access
token ceased to work for dispatching to the functional test on
deploys.  We still want that to happen, so I made a token with the
`public_repo` scope (as suggested by the
[documentation](https://github.com/marketplace/actions/repository-dispatch#token),
since the functional test is a public repo) and saved that as a secret
on this repository.  If and when I am removed from this organization
the token will need to be replaced again.