Run database migrations only during code deploy
Simplify the work of deployments by automatically running database
migrations, rather than requiring an engineer to manually do so. This
removes the risk of the code being ahead of the database structure
long-term, reduces the window between code being deployed and the
database looking as the code expects, and reduces deployment time
overall.

We were briefly running database migrations automatically as part of
deploying[1], but ran into a problem and turned automatic migrations off[2].
The problem we encountered was that the image builder instance was not
always on the same subnet as the production instance, and so sometime
image building would fail because migrations could not be run. That shed
light on a more important problem, though: we need to not run migrations
during image creation! There can be a substantial delay between building
an image and running terraform to bring up new instances based on that
image, and during that time the database can be ahead of the code.

Instead, we need to run database migrations only during code deploys,
and not during image creation. Reintroduce the migrations task[1], and
add new variable that we only set during the GitHub Actions workflows
for code deployments, and not during the image building workflow. Use an
Ansible conditional[3] to skip the migration task during image
building.

[1] commit 4ab20a62b0860107d7a1c5723a161582af9b0954
[2] commit 535eb99ee78edc5fb15f45d939b8183850f182d8
[3] https://docs.ansible.com/ansible/latest/user_guide/playbooks_conditionals.html#conditionals-based-on-variables

PER-8780 Run database migrations on code deployment

Co-authored-by: Cecilia Krum <cecilia@permanent.org>