Use pinned apt repo signing keys and fix MySQL
We use a handful of third-party Debian repos. We were using apt-key(8)
to install those repos' signing keys globally, which was the style at
the time, but is no longer. From the man page:

     Use of apt-key is deprecated, except for the use of apt-key del in
     maintainer scripts to remove existing keys from the main keyring.

Instead, the key(s) for the repository should be stored in
/usr/share/keyrings, and the apt source should specify that keyring.

See also:
- https://www.linuxuprising.com/2021/01/apt-key-is-deprecated-how-to-add.html
- https://askubuntu.com/a/1307181
- https://unix.stackexchange.com/a/582853

Alone, that is a great bit of future-proofing, and an improvement to our
security posture, but the real motivation for these changes is to update
the MySQL signing key.

A few weeks ago, Oracle started signing their MySQL apt repo with a new
key[1], partially in response to the old key expiring[2]. While we only
use the MySQL repo in production for the mysql command line client (and
that only for the convenience of devs), the key rotation broke the
ability for devs to build new working Vagrant VMs.

Save the third-party apt signing keys, update our apt sources to use
them, and update our provisioning scripts to install both.

The signing keys were taken from the respective third party
repositories:
- https://deb.nodesource.com/gpgkey/nodesource.gpg.key
- https://download.newrelic.com/548C16BF.gpg
- https://repo.mysql.com/RPM-GPG-KEY-mysql-2022

Note that the signing keys are all in their armored plaintext form, as
distributed by upstream. This is both to allow for easy comparison to /
verification of the upstream keys, and because binary files are not
great in git. It seems that apt support for armored keys has been around
since Debian stretch[3], and was added after support for binary keys,
and so my understanding is that advice to always dearmor keys is legacy
knowledge intended to maintain backward compatibility with a release we
are no longer using and that will shortly be out of extended support.

Thanks to @slifty for doing some of this research already, in
torque-devenv[4].

[1] https://bugs.mysql.com/bug.php?id=106188
[2] https://bugs.mysql.com/bug.php?id=105632
[3] https://wiki.debian.org/DebianRepository/UseThirdParty
[4] https://github.com/OpenTechStrategies/torque-devenv/commit/6e01a7e51818ec60e538124b2bdea0c673245895