Install packages all at once
Previously, we were installing apt packages in several passes. Part of
the motivation was to make available the tools we needed to configure
third party repos, like curl and gpg. Now that we are more securely
hardcoding those repos[1], rather than fetching configuration over the
internet, those tools don't need to be installed in a separate early
pass.

There are a few drawbacks to installing in multiple passes. A small one
is time: invoking apt multiple times takes longer than invoking it only
once, which leads to longer build times.

A bigger problem is it makes it a bit harder for devs to know where and
how to add a package to the install list: should it be in one of the
existing installation steps? where in the list should it be added?
should the existing comment be updated? should it be installed in its
own install step? should that have a comment?

Finally, and perhaps most importantly, it makes it difficult to audit
the packages we're installing, as that information needs to be drawn
from multiple places.

Consolidate all the package installation steps into a single step. Split
the packages out to one line per package, and sort them alphabetically.

[1] see commit bc511b890fb4f4c6d82f1310f651719e57b7f9c4