Add PostgreSQL client from official repo
Configure the official PostgreSQL apt repo, and install the client so
that developers can use the provisioned machines to connect to the
database.

The official repo is needed to use the current version of PostgreSQL.
Debian does provide PostgreSQL packages, but of course the Debian and
PostgreSQL release schedules do not perfectly align, and anyway we're
still on oldstable. We'll be hosting our production databases in Amazon
RDS, so we don't need or want to use the Debian-provided server version.
Problems arise if you try to connect to a newer server with an older
client, so it's worthwhile to make the versions match everywhere.

This will also help set up the local devenv, as it will need the apt
repo configured to install the server.

The repository key is from the website[1], as specified on the wiki[2].

[1] https://www.postgresql.org/media/keys/ACCC4CF8.asc
[2] https://wiki.postgresql.org/wiki/Apt

PER-8846 Set up MySQL->PostgreSQL switch