PER-9158: Deeplink to reset password and signup
Add well known links for resetting forgotten passwords and signing up
to Permanent. Note if the reset password link location changes in
https://github.com/PermanentOrg/web-app/pull/192/files then it will
need to change here to match.

I'm not sure if /* is necessary or if the default is to match on
substrings. The reset password links will have some key strings at the
end to indicate the request and the FusionAuth tenant. Signup urls
could end with "signup" or could have a querystring on the end.