Increase sftp storage to 1TB
In the short term we would like to mitigate the risk of the sftp service
getting filled up by temporary files.  1TB doesn't prevent the risk
entirely, but it does make it harder for the drive to get filled by
a single user uploading a whole lot of content at once.