Use the updated version of node to run upload-service
When we install our desired version of `node` on the backend
instance, it can be invoked with the command: `node`. Currently,
`upload-service` is started with `/usr/bin/nodejs`, which is not
necessarily our desired version. This commit updates the
`upload-service` template to use `node` instead, so it will run on the
desired node version.