PER-9263: Basic web-app routing for WJMA
Redirecting a specific page to the webapp as we've done
elsewhere. This is a partial step. We've discussed handling the entire
redirection in this file, rather than also having a redirect line in
the Angular routing.

Also interested in whether reviewers think it makes sense to have this
across all configuration files, for consistency, or only on prod since
that's the only place where the redirection will be useful (in other
environments it'll most likely reach a page that says the archive
doesn't exist).