Add service to clean dead files
The SFTP service will use temporary storage for files in transit.

There are various ways that a temporary file might not be properly
cleaned up (e.g. if the SFTP process unexpectedly fails).  To prevent
a "storage leak" we want to regularly delete any temporary files that
have not been modified for several days.