Use current stable version of php package
Debian provides a dependency package for php, "which depends on latest
stable PHP version". We are using similar such dependency packages for
all the php extensions, which each depend on the version-specific
package. Newer versions of Debian continue to provide dependency
packages with the same names, but which depend on the version of php
packaged in that version of Debian; this makes it easier to upgrade
Debian, as the same package installation command will continue to work.

Use the dependency package for php rather than the version-specific
package.

Issue #75 Upgrade Debian
Issue #130 Configure Apache to use mpm_event