Switch from mpm_prefork to mpm_event
Use Apache's event multi-processing module[1], which

    is designed to allow more requests to be served simultaneously by
    passing off some processing work to the listeners threads, freeing
    up the worker threads to serve new requests.

This should allow us to better handle concurrent requests, as are
sometimes created by the sftp-service.

An additional benefit of switching away from mpm_prefork is that we can
now enable HTTP/2 via the http2 module, which mpm_prefork conflicts
with. The AWS load balancer already handles HTTP/2 requests from
clients, but this allows for the load balancer to use HTTP/2 in its
connections to Apache, which should improve performance.

We no longer need to specify the `Mutex` directive[2], as mpm_event does
not require it:

    This new architecture, leveraging non-blocking sockets and modern
    kernel features exposed by APR (like Linux's epoll), no longer
    requires the mpm-accept Mutex configured to avoid the thundering
    herd problem.

Remove it from apache2.conf to bring it closer to the standard version
packaged by Debian.

[1] https://httpd.apache.org/docs/2.4/mod/event.html
[2] https://httpd.apache.org/docs/2.4/mod/core.html#mutex

Issue #130 Configure Apache to use mpm_event