Put Apache configs into sites-available
The way Apache expects to be configured is to have a number of config
files with `<VirtualHost>` directives in `/etc/apache2/sites-available`,
and then for `/etc/apache2/sites-enabled` to contain symlinks to those
sites. The closer we are to the standard way of configuring Apache, the
easier it'll be for engineers to work with the Apache configuration!