Move global ServerName directive to its own file
Omitting a global `ServerName` directive in the Apache configuration can
lead to warnings in the logs:

    AH00558: apache2: Could not reliably determine the server's fully
    qualified domain name, using 127.0.0.1. Set the 'ServerName'
    directive globally to suppress this message

We want to use the stock Debian-packaged Apache configuration file, so
move this bit of configuration out of `apache2.conf` into a file in
`conf-available` and then enable it. This both makes it easier to
upgrade Debian versions and makes it easier for engineers familiar with
Apache to understand how the system is configured.

Issue #75 Upgrade Debian