Tune mpm_event and php_fpm
The default settings for mpm_event and php_fpm are very conservative;
for example, the stock configuration file
/etc/php/7.3/fpm/pool.d/www.conf calls this out:

    The below defaults are based on a server without much resources.

The sole purpose of the API server is to serve the API, so all the
available resources can and should be dedicated to that.

Raise the number of Apache and PHP threads each. Raising them too high
would result in out-of-memory errors, so these are still somewhat
conservative numbers. I was able to double them on dev and successfully
run my synthetic load test, but proper performance tuning should involve
a detailed and representative load test and take advantage of the
existing performance measurements we are collecting via New Relic.

Modify the php_fpm configuration via a patch file rather than a template
file in order to ease upgrading Debian - if the new version of the
php_fpm package changes other, unrelated settings in the default
configuration, we'll want to pick them up, rather than silently
overwriting them. Unfortunately, php_fpm does not have as sophisticated
a configuration system as Apache, so we do have to modify in place.

Issue #75 Upgrade Debian
Issue #130 Configure Apache to use mpm_event