Increase api image size
As part of our work exploring the impact of the sftp service, we
identified the need to increase the perfomance capability of the API
server in order to handle the volume of API calls made by SFTP users
[1].

Amazon provides a range of instance sizes.  This choice increases the
CPU capacity of the machine without expanding the memory capacity.  This
is because our initial benchmarks seemed to be pinning CPU usage but not
putting a particularly large dent in memory.

[1] https://github.com/PermanentOrg/sftp-service/issues/268