Add new sftp environment variables
The SFTP service needs some additional environment variables to be
populated, as a result of some recent changes to how refresh tokens are
used to generate auth tokens [1].

A few of these vriables are redundant [2], and that's why we use the
same "source" variable to map them as late as possible in the
provisioning.  Eventually if the redundancy is removed from the sftp
service we'll want to update the provisioner to stop populating the
obsolete copies.

Issue #138

[1] https://github.com/PermanentOrg/sftp-service/pull/192
[2] https://github.com/PermanentOrg/sftp-service/issues/289