Clean Fusion Auth environment variables
With the latest changes to the sftp service we have removed some
redundancy in the configuration variables. This commit responds to that
cleanup, while also making some corrections to some observed
inconsistencies in the naming of certain environment variables AND the
consistency in referencing them in certain pieces of the code (e.g. the
provisioning script).

Issue #140 Adjust SFTP environment variables