Add SSL certificates for AWS RDS to deployed instances
Presently, we don't use SSL in our database connections. This is a
security best practice, so we should. This commit adds the public key
certificates for the us-west-2 RDS signing authorities to the images we deploy,
so they can be specified as trusted signing authorities when we connect to the
database.