Update nodesource gpg keys
Nodesource updated their GPG keys and switched to using `nodistro`

This is a second attempt to reflect those changes.  It should NOT be
merged without first understanding what else has to happen for this to
work in our image building, at which point this commit message should be
updated.

Issue #178 Figure out what's up with nodesource