Fix patch pathing
Our patch is being applied with a directory already in our configuration
script, making the definition inside of the patch redundant (and also,
as we have seen, a potential source of failure!)

When we upgraded to php 8.3 the patch path wasn't updated and so
configuration tweaks were being lost.

This change should help prevent that from happening in future updates as
well.

Issue #190 Fix our "patch" of php-fpm