Add php-fpm timeout
We noticed in some cases a given php-fpm thread can lock while waiting
on a read.  This lock was indefinite, and was resulting in a cascading
deadlock that translated to a service outage.

We want to kill any php process that gets into this state.  We felt that
60 seconds was a reasonable amount of time.  I believe php itself has
a max runtime of 30 seconds so this should really only trigger in rare
cases.

Issue #190 Fix our "patch" of php-fpm