Increase PHP memory limit
Currently, some users of permanent have folders large enough that
retrieving all their contents via the API causes PHP to run out of
memory. The correct way to fix this is to paginate those API calls, but
as a hotfix, this commit increases the memory limit so users will still
be able to access their files.

Since this commit adds a patch for the php.ini config file, it also
includes the update to gc_maxlifetime that we've historically made
manually on every deploy.